# probe-api/api/openapi.yaml
openapi: 3.1.0
info:
  title: Probe Panel API
  version: 1.0.0-draft
  summary: Frozen V1 contract for the panel, administrators, Agents, optional public API, and health checks.
  description: |
    The visitor panel and administrator panel are independent static applications and origins.
    `panel.example.com` serves only probe-web and anonymous read-only `/api/v1/panel/*` routes.
    `admin.example.com` serves only probe-admin, administrator `/api/v1/auth/*` and `/api/v1/admin/*`
    routes, plus the same read-only `/api/v1/panel/*` data needed by management views. Both browser
    origins remain behind the management IP/CIDR allowlist; anonymous does not mean public Internet
    access. Only administrators have accounts, passwords, and sessions. Agent and optional public
    routes are exposed only on `api.example.com`. Host/path exposure,
    management IP/CIDR allowlisting, request size limits, Origin checks, and rate limits are enforced
    before or together with the authentication described here. Unlisted paths and methods are denied.
    All timestamps are RFC 3339 instants and all server-generated timestamps use UTC.
    This document also describes the short-lived first-run Setup API. Every Setup path keeps the
    local-browser URL `http://127.0.0.1:18080`, but the server listens only on the root-owned Unix
    socket `/run/probe-panel-setup/setup.sock`. Reach it while the machine is uninitialized through
    a root SSH Unix-socket forward such as
...
servers:
...
security: []
paths:
  /api/v1/setup/status:
    servers:
      - url: http://127.0.0.1:18080
        description: Local-browser URL forwarded by root SSH to `/run/probe-panel-setup/setup.sock`; no remote TCP listener exists.
    get:
      tags: [setup]
      operationId: getSetupStatus
      summary: Read the current first-run installation state.
      description: |
        This route exists only on the root-owned `probe-setup` Unix socket while first-run setup is
        active. A local browser reaches it through a root SSH Unix-socket forward. `Host` must be
        exactly `127.0.0.1:18080` or `localhost:18080`. `Origin` may be omitted by a non-browser GET;
        when present it must be exactly `http://127.0.0.1:18080` or `http://localhost:18080`.
        An `installed` response may briefly include the administrator login URL and, for IP mode,
        the public private-CA certificate plus its exact whole-file SHA-256 before the temporary
        service shuts down. The browser must retain that CA only in page memory. Production Nginx
        does not proxy this route and permanently returns 404.
      x-access-boundary: temporary-root-unix-socket-ssh-forward-only
      security: []
      responses:
        '200':
          description: Current setup state; terminal handoff metadata may be present only while the temporary service reports `installed`.
          headers:
            Cache-Control:
              schema: { type: string, const: no-store }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SetupStatusResponse' }
        '403': { $ref: '#/components/responses/SetupForbidden' }
        '503': { $ref: '#/components/responses/SetupUnavailable' }
  /api/v1/setup/session:
    servers:
      - url: http://127.0.0.1:18080
        description: Local-browser URL forwarded by root SSH to `/run/probe-panel-setup/setup.sock`; no remote TCP listener exists.
    post:
      tags: [setup]
      operationId: createSetupSession
      summary: Mint or rotate the sole short-lived in-memory setup session.
      description: |
        Accepts only the strict empty JSON object `{}` over the root-owned Unix socket and returns a
        new 15-minute session/CSRF pair held only in memory. The first request advances `pending` to
        `configuring`; a later request while configuring invalidates the previous pair and mints a
        replacement, so a browser refresh or service restart can recover by rotating credentials. The body is limited to
        64 KiB; properties, arrays, duplicate/trailing JSON, and non-empty bodies are rejected. The
        `Host` restrictions are the same as the status route and the exact allowed `Origin` is
        mandatory.
      x-access-boundary: temporary-root-unix-socket-ssh-forward-origin
      security: []
      parameters:
        - $ref: '#/components/parameters/SetupOrigin'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetupSessionRequest' }
      responses:
        '200':
          description: The previous pair, if any, was invalidated and one short-lived setup session was created.
          headers:
            Cache-Control:
              schema: { type: string, const: no-store }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SetupSessionResponse' }
        '400': { $ref: '#/components/responses/SetupBadRequest' }
        '403': { $ref: '#/components/responses/SetupForbidden' }
        '409': { $ref: '#/components/responses/SetupConflict' }
        '413': { $ref: '#/components/responses/SetupPayloadTooLarge' }
        '415': { $ref: '#/components/responses/SetupUnsupportedMediaType' }
        '429':
          description: More than five setup-session mint requests were made in the current one-minute window.
          headers:
            Retry-After:
              schema: { type: integer, minimum: 1 }
            Cache-Control:
              schema: { type: string, const: no-store }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SetupErrorResponse' }
  /api/v1/setup/complete:
    servers:
      - url: http://127.0.0.1:18080
        description: Local-browser URL forwarded by root SSH to `/run/probe-panel-setup/setup.sock`; no remote TCP listener exists.
    post:
      tags: [setup]
      operationId: completeSetup
      summary: Validate the first-run configuration and start privileged finalization asynchronously.
      description: |
        Requires the exact `X-Probe-Setup-Session` and `X-CSRF-Token` values returned by the session
        exchange; `Authorization: Bearer` is not accepted. It also requires an allowed `Origin`, an
        allowed loopback `Host`, the authenticated root-owned Unix-socket transport, and an `application/json` body no larger than
        64 KiB. The JSON object is strict: unknown, duplicate, missing, and trailing fields are
        rejected. Password confirmations are compared in constant time. Accepted secrets are copied
        only for the asynchronous finalizer and are cleared when it returns. Production Nginx never
        exposes this path.
      x-access-boundary: temporary-root-unix-socket-ssh-forward-session-origin-csrf
      security:
        - SetupSession: []
          SetupCsrfToken: []
      parameters:
        - $ref: '#/components/parameters/SetupOrigin'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetupCompleteRequest' }
      responses:
        '202':
          description: Configuration validation passed and finalization started; poll the status route.
          headers:
            Cache-Control:
              schema: { type: string, const: no-store }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SetupFinalizingResponse' }
        '400': { $ref: '#/components/responses/SetupBadRequest' }
        '401': { $ref: '#/components/responses/SetupUnauthorized' }
        '403': { $ref: '#/components/responses/SetupForbidden' }
        '409': { $ref: '#/components/responses/SetupConflict' }
        '413': { $ref: '#/components/responses/SetupPayloadTooLarge' }
        '415': { $ref: '#/components/responses/SetupUnsupportedMediaType' }
...
components:
  securitySchemes:
    SetupSession:
      type: apiKey
      in: header
      name: X-Probe-Setup-Session
      description: One 64-character lowercase hexadecimal token minted over the root-only setup Unix socket and held only in memory. `Authorization` is not accepted.
    SetupCsrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: One 64-character lowercase hexadecimal token bound to the in-memory setup session; unrelated to administrator-session CSRF tokens.
    SessionCookie:
      type: apiKey
      in: cookie
      name: probe_session
      description: Host-only HttpOnly Secure SameSite=Strict administrator session token. V1 defaults to a 12-hour absolute lifetime and five concurrent sessions per administrator.
    CsrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Stable per-session token derived with domain-separated HMAC-SHA256 and returned by login and auth/me; required with exact Origin on cookie-authenticated writes. Repeated auth/me calls do not rotate it.
    AgentBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-agent-token
    PublicApiKey:
      type: apiKey
      in: header
      name: X-Probe-API-Key
      description: Independent key for the disabled-by-default public read-only API.
    InternalBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-internal-token
      description: Required only when an internal endpoint is deliberately exposed beyond loopback.
  parameters:
    SetupOrigin:
      name: Origin
      in: header
      required: true
      description: Exact local-browser origin accepted after root SSH forwards traffic to the setup Unix socket.
      schema:
        type: string
        enum:
          - http://127.0.0.1:18080
          - http://localhost:18080
...
  responses:
    SetupBadRequest:
      description: The strict setup JSON document is malformed or fails field validation.
      headers:
        Cache-Control:
          schema: { type: string, const: no-store }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/SetupErrorResponse' }
...
    SetupForbidden:
      description: The root Unix-socket transport, exact local Host, exact Origin, or setup CSRF check failed.
      headers:
        Cache-Control:
          schema: { type: string, const: no-store }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/SetupErrorResponse' }
    SetupConflict:
      description: The persistent setup state does not allow the requested transition or finalization is already active.
      headers:
        Cache-Control:
          schema: { type: string, const: no-store }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/SetupErrorResponse' }
    SetupPayloadTooLarge:
      description: The setup JSON request body exceeds 64 KiB.
      headers:
        Cache-Control:
          schema: { type: string, const: no-store }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/SetupErrorResponse' }
    SetupUnsupportedMediaType:
      description: The request Content-Type is not application/json.
      headers:
        Cache-Control:
          schema: { type: string, const: no-store }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/SetupErrorResponse' }
    SetupUnavailable:
      description: Root-owned setup state could not be read safely.
      headers:
        Cache-Control:
          schema: { type: string, const: no-store }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/SetupErrorResponse' }
...
  schemas:
    SetupDefaults:
      $ref: '#/components/schemas/SetupManagementDefaults'
...
    SetupStatusResponse:
      description: |
        `pending` is ready to mint a setup session; `configuring` holds the one currently valid
        in-memory session or indicates that a side-effect-free privileged preflight was rejected and
        a new session may be created; `finalizing` means the privileged asynchronous finalizer is
        running; `installed` is terminal success; `recovery_required` is reserved for a failure after
        production ownership/mutation and requires operator recovery rather than reopening setup.
      oneOf:
        - type: object
          additionalProperties: false
          required: [status, defaults]
          properties:
            status:
              type: string
              enum: [pending, configuring, finalizing, recovery_required]
            defaults: { $ref: '#/components/schemas/SetupDefaults' }
        - type: object
          additionalProperties: false
          required: [status, defaults]
          properties:
            status: { type: string, const: installed }
            defaults: { $ref: '#/components/schemas/SetupDefaults' }
            admin_url:
              type: string
              format: uri
              pattern: '^https://[^/]+/login$'
              description: Optional administrator login URL constructed from the installed domain or IP origin.
            private_ca:
              $ref: '#/components/schemas/SetupPrivateCAHandoff'
            handoff_unavailable:
              type: boolean
              const: true
              description: The persisted installation succeeded, but safe terminal handoff metadata could not be reconstructed; the UI must show the root SSH/scp fallback and must not redirect automatically.
      example:
        status: installed
        defaults:
          profile: management
          server_ip: 192.0.2.10
          admin_url: https://192.0.2.10:18455
        admin_url: https://192.0.2.10:18455/login
        private_ca:
          available: true
          pem: |
            -----BEGIN CERTIFICATE-----
            BASE64_CERTIFICATE_DATA
            -----END CERTIFICATE-----
          sha256: 0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef
...
    SetupSessionRequest:
      type: object
      additionalProperties: false
      maxProperties: 0
      description: Must be exactly the empty JSON object `{}`; the root-only Unix socket is the authorization boundary.
    SetupSessionResponse:
      type: object
      additionalProperties: false
      required: [session_token, csrf_token, expires_at, defaults]
      properties:
        session_token:
          type: string
          minLength: 64
          maxLength: 64
          pattern: '^[0-9a-f]{64}$'
          readOnly: true
        csrf_token:
          type: string
          minLength: 64
          maxLength: 64
          pattern: '^[0-9a-f]{64}$'
          readOnly: true
        expires_at:
          type: string
          format: date-time
          readOnly: true
          description: RFC 3339 expiry of this single in-memory session, normally 15 minutes after exchange.
        defaults: { $ref: '#/components/schemas/SetupDefaults' }
...